backend/requirements.txt
Django==4.1.13  # Downgrading to 4.1.x which is compatible with django-tenant-users
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-tenant-users==1.1.1
django-cors-headers==4.0.0
celery==5.2.7
redis==4.5.5
python-decouple==3.8
Pillow==9.5.0
django-filter==23.3
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
dj-database-url==2.0.0
gunicorn==20.1.0
plaid-python==16.0.0
